refactor: Extract write functionality
